auth v1

+++ application/controllers/api/Log_Aktivitas.php
@@ -13,6 +13,18 @@ class Log_Aktivitas extends CI_Controller{
         parent::__construct();
         $this->__resTraitConstruct();
         $this->load->model('Log_Aktivitas_model','akt');
+        $this->load->model('MyModel','mm');
+        // $check_auth_client = $this->mm->check_auth_client();
+        // if($check_auth_client == true){
+        //     $response = $this->mm->auth();
+        // if($response['status'] == 200){
+        // json_output($response['status']);
+        // } else {
+        // json_output(400,array(
+        //     'status' => 400,
+        //     'message' => 'Bad request.'));
+        //     }
+        // }
     }
     //-----------------------------------------------------------------------------------------//
                                             //Method GET//
@@ -24,33 +36,17 @@ class Log_Aktivitas extends CI_Controller{
         $month = $this->get('month');
         $method = $_SERVER['REQUEST_METHOD'];
         if($method != 'GET'){
-        json_output(400,array('status' => 400,'message' => 'Bad request.'));
+			json_output(400,array('status' => 400,'message' => 'Bad request.'));
+		} else {
+            $check_auth_client = $this->mm->check_auth_client();
+			if($check_auth_client == true){
+		        $response = $this->mm->auth();
+        if($response['status'] == 200 && $nip != null){
+            $akt = $this->akt->getAktivitas($nip,$year,$month);
+            json_output($response['status'],$akt);
         } else {
-                    $check_auth_client = $this->MyModel->check_auth_client();
-        if($check_auth_client == true){
-                $response = $this->MyModel->auth();
-                if($nip != null){
-                    $akt = $this->akt->getAktivitas($nip,$year,$month);
-                } else {
-                    $akt = $this->akt->getAktivitas();
-                }
-                
-                if($akt){
-                    $this->response([
-                        'status' => true,
-                        'data'   => $akt
-                    ], 200);
-                } else{
-                    $this->response([
-                        'status'  => false,
-                        'message' => 'Maaf, ID tidak ditemukan !'
-                    ], 404);
-                }
-            }
+            $akt = $this->akt->getAktivitas();
         }
-    }
-    public function aktifitas_get() {
-        $akt = $this->akt->getAktivitasAll();
         
         if($akt){
             $this->response([
@@ -62,6 +58,34 @@ class Log_Aktivitas extends CI_Controller{
                 'status'  => false,
                 'message' => 'Maaf, ID tidak ditemukan !'
             ], 404);
+                }
+            }
+        }
+    }
+    public function aktifitas_get() {
+        $method = $_SERVER['REQUEST_METHOD'];
+        if($method != 'GET'){
+			json_output(400,array('status' => 400,'message' => 'Bad request.'));
+		} else {
+            $check_auth_client = $this->mm->check_auth_client();
+			if($check_auth_client == true){
+		        $response = $this->mm->auth();
+        if($response['status'] == 200){
+            $akt = $this->akt->getAktivitasAll();
+            json_output($response['status'],$akt);
+        }
+        if($akt){
+            $this->response([
+                'status' => true,
+                'data'   => $akt
+            ], 200);
+        } else{
+            $this->response([
+                'status'  => false,
+                'message' => 'Maaf, ID tidak ditemukan !'
+            ], 404);
+                }
+            }
         }
     }
     //-----------------------------------------------------------------------------------------//
@@ -71,8 +95,17 @@ class Log_Aktivitas extends CI_Controller{
     public function before_get() {
         $nip = $this->get('nip');
         $before = $this->get('before');
-
-        $akt = $this->akt->getTanggalBefore($nip,$before);
+        $method = $_SERVER['REQUEST_METHOD'];
+        if($method != 'GET'){
+			json_output(400,array('status' => 400,'message' => 'Bad request.'));
+		} else {
+            $check_auth_client = $this->mm->check_auth_client();
+			if($check_auth_client == true){
+		        $response = $this->mm->auth();
+        if($response['status'] == 200 && $nip != null){
+            $akt = $this->akt->getTanggalBefore($nip,$before);
+            json_output($response['status'],$akt);
+        }       
         
         if($akt){
             $this->response([
@@ -84,12 +117,24 @@ class Log_Aktivitas extends CI_Controller{
                 'status'  => false,
                 'message' => 'Maaf, ID tidak ditemukan !'
             ], 404);
+                }
+            }
         }
     }
     public function after_get() {
         $nip = $this->get('nip');
         $after = $this->get('after');
-        $akt = $this->akt->getTanggalAfter($nip,$after);
+        $method = $_SERVER['REQUEST_METHOD'];
+        if($method != 'GET'){
+			json_output(400,array('status' => 400,'message' => 'Bad request.'));
+		} else {
+            $check_auth_client = $this->mm->check_auth_client();
+			if($check_auth_client == true){
+		        $response = $this->mm->auth();
+        if($response['status'] == 200  && $nip != null ){
+            $akt = $this->akt->getTanggalAfter($nip,$after);
+            json_output($response['status'],$akt);
+        }
         
         if($akt){
             $this->response([
@@ -101,13 +146,25 @@ class Log_Aktivitas extends CI_Controller{
                 'status'  => false,
                 'message' => 'Maaf, ID tidak ditemukan !'
             ], 404);
+                }
+            }
         }
     }
     public function between_get() {
         $nip = $this->get('nip');
         $before = $this->get('before');
         $after = $this->get('after');
-        $akt = $this->akt->getTanggalBetween($nip,$before,$after);
+        $method = $_SERVER['REQUEST_METHOD'];
+        if($method != 'GET'){
+			json_output(400,array('status' => 400,'message' => 'Bad request.'));
+		} else {
+            $check_auth_client = $this->mm->check_auth_client();
+			if($check_auth_client == true){
+		        $response = $this->mm->auth();
+        if($response['status'] == 200  && $nip != null ){
+            $akt = $this->akt->getTanggalBetween($nip,$before,$after);
+            json_output($response['status'],$akt);
+        }
         
         if($akt){
             $this->response([
@@ -119,6 +176,8 @@ class Log_Aktivitas extends CI_Controller{
                 'status'  => false,
                 'message' => 'Maaf, ID tidak ditemukan !'
             ], 404);
+                }
+            }
         }
     }
 
@@ -127,8 +186,18 @@ class Log_Aktivitas extends CI_Controller{
                                           //DELETE by log_id//
     //-----------------------------------------------------------------------------------------//
     public function index_delete() {
-        $log_id = $this->delete('log_id');
-
+        $method = $_SERVER['REQUEST_METHOD'];
+        if($method != 'GET'){
+			json_output(400,array('status' => 400,'message' => 'Bad request.'));
+		} else {
+            $check_auth_client = $this->mm->check_auth_client();
+			if($check_auth_client == true){
+		        $response = $this->mm->auth();
+        if($response['status'] == 200  && $log_id != null ){
+            $log_id = $this->delete('log_id');
+            json_output($response['status'],$akt);
+        }
+        
         if($log_id === null){
             $this->response([
                 'status' => false,
@@ -147,22 +216,33 @@ class Log_Aktivitas extends CI_Controller{
                     'status' => false,
                     'data'   => 'Maaf, ID tidak ditemukan !'
                 ], 400);
-            }
+                    }
+                }
+            }   
         }
     }
     //-----------------------------------------------------------------------------------------//
                                             //Method POST//
     //-----------------------------------------------------------------------------------------//
     public function index_post() {
-        $data = [
-            'log_id'        => $this->post('log_id'),
-            'akt_tanggal'   => $this->post('akt_tanggal'),
-            'akt_idkegiatan'=> $this->post('akt_idkegiatan'),
-            'akt_catatan'   => $this->post('akt_catatan'),
-            'akt_output'    => $this->post('akt_output'),
-            'akt_start'     => $this->post('akt_start'),
-            'akt_end'       => $this->post('akt_end')
-        ];
+        $method = $_SERVER['REQUEST_METHOD'];
+		if($method != 'POST'){
+			json_output(400,array('status' => 400,'message' => 'Bad request.'));
+		} else {
+			$check_auth_client = $this->MyModel->check_auth_client();
+			if($check_auth_client == true){
+		        $response = $this->MyModel->auth();
+		        $respStatus = $response['status'];
+		        if($response['status'] == 200){
+                    $data = [
+                        'log_id'        => $this->post('log_id'),
+                        'akt_tanggal'   => $this->post('akt_tanggal'),
+                        'akt_idkegiatan'=> $this->post('akt_idkegiatan'),
+                        'akt_catatan'   => $this->post('akt_catatan'),
+                        'akt_output'    => $this->post('akt_output'),
+                        'akt_start'     => $this->post('akt_start'),
+                        'akt_end'       => $this->post('akt_end')
+                    ];
         if ($this->akt->createAktivitas($data) > 0) {
             $this->response([
                 'status'  => true,
@@ -173,8 +253,11 @@ class Log_Aktivitas extends CI_Controller{
                 'status' => false,
                 'data'   => 'Maaf, Data baru gagal dibuat !'
             ], 400);
+                        }
+                    }
+                }
+            }
         }
-    }
     //-----------------------------------------------------------------------------------------//
                                             //Method PUT//
                                         //EDIT data by log_id//
